refactor nodes, implement new search
