+++ html/pfappserver/root/src/views/Nodes/_composables/useCollection.js
@@ -1,2 +1,3 @@
 export { useRouter } from '../_router'
+export { useSearch } from '../_search'
 export { useStore } from '../_store'
